+++ HSYCocoaKit/Classes/OpenSource/RSA/RSA.m
@@ -262,7 +262,7 @@ static NSData *base64_decode(NSString *str){
                                    );
         }
 		if (status != 0) {
-			NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
 			ret = nil;
 			break;
 		}else{
@@ -318,7 +318,7 @@ static NSData *base64_decode(NSString *str){
 							   &outlen
 							   );
 		if (status != 0) {
-			NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
 			ret = nil;
 			break;
 		}else{
